integrate land-sea masking into file_conversion
